+++ integration/tarantoolee_gost_rsa_clientauth_test.go
@@ -33,8 +33,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	tlsdialer "github.com/tarantool/go-tlsdialer"
-	"github.com/tarantool/go-tlsdialer/backend/gostls"
+	tlsdialer "github.com/tarantool/go-tlsdialer/v2"
+	"github.com/tarantool/go-tlsdialer/v2/backend/gostls"
 )
 
 func TestTarantoolEE_Ping_GOSTServer_RSAClientAuth(t *testing.T) {